Rewrite the parsing system supporting more logger
Instead using a chain of try and catches an intelligent, at least a
combined, parser is used which checks if the parser can handle the line
and if, the parser is executed.
In further update the intelligent parser will reorder the call order of
the parser to improve the speed of parsing by calling the parser with the
highest chance to succeed.